requirements.txt
accelerate==0.27.2
datasets==2.17.1
evaluate==0.4.1
fastchat==0.1.0
fschat==0.2.36
huggingface_hub==0.21.3
hydra-core==1.3.2
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
pandas==2.2.2
peft==0.9.0
tinydb==4.8.0
torch==2.1.2
tqdm==4.66.2
transformers==4.41.3
trl==0.8.3